LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/scroll-initial-target/scroll-initial-target-span.tentative-expected.txt


FAIL <span> scroll-initial-target is honored assert_equals: scroller is initially scrolled to it <span> scroll-initial-target expected 128 but got 0
